pyproject.toml
[tool.poetry]
name = "azure-devops-kit"
version = "0.1.0"
description = "Azure DevOps Kit for test case management and analysis"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.3"
azure-devops = "^6.0.0b4"
pandas = "^2.0.0"
matplotlib = "^3.7.1"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
azdo-kit = "azure_devops_kit.cli:main"
